Check for needed keys with values of the correct format in the uploaded file using Yup
