+++ frontend/README.md
@@ -1,4 +1,4 @@
 # Getting Started
 
 1. Run `npm install` inside the frontend -folder to install dependencies. This needs to be run only once when you clone the project.
-2. Run `npm start` (inside the folder) to start the app.
+2. Run `npm start` (inside the frontend folder) to start the app.
